Move Create_json function to util.py
